make files structure more flat; move ./CodeSamples to ./Download
